Add endpoints to advance tutorial (#15)
Adds:

- /friend/get_support_chara (stub)
- /information/top (used once for a joke)
- /mission/get_mission_list (stub)
- /party/set_party_setting

Also adds a news Razor page to replace the in-game news, but it is
non-functional in dev environments, as the game will not make requests
to the localhost:5000 URL that is sent by /webview_version/get_url_list.
This should work if the server is deployed to a dedicated domain though.