Migrate to PostgreSQL (#33)
Migrating to PostgreSQL from SQL Server due to licensing concerns.

Additionally reconfigures the way database passwords are handled: 
- Previously these were in appsettings.json and the docker-compose.yml
file, which not only left them in plaintext and presented a security
vulnerability if unchanged, but also presented two sources of truth.
- With this PR, they are now derived from environment variables
specified in postgres.env which is left blank by default. This file is
in the .gitignore so that database password will not accidentally be
committed.