Add support for Kaleidoscape wyrmprints obtained via savefile import (#41)
Adds a new database entity for tracking Kaleidoscape wyrmprints
(talismans). This table will now be read from on loading a savefile and
written to when importing a savefile. Units will also have talisman data
pulled from the key_id. when entering a quest.

There is a similar constraint on the key_id as with dragons that all
existing key_ids in parties are set to 0, thus talismans must be
re-equipped after importing a savefile.

There is currently no other way to get talismans other than importing a
savefile.

Closes #35