Shared skill fix (#65)
Closes #64 

Story characters will have their shared skills unlocked by default when
being added (not backwards compatible)

Weapon skills and abilities are always set to the max level but this
will need to be revisited when weapon upgrading is implemented. These
values are not tracked in the database because they are not sent by
/load/index, so presumably they're calculated on the fly. Additionally,
skill_no is always set to 1 because I am unsure what this does, however
it appears to work with Agito/PDT weapons just fine.