Data service rework (#70)
Reworks the various data services to not be distinct types, but instead
implementations of a base service which is composed with a
KeyedCollection. Full JSON files are used with the models containing the
subset of fields that we actually need. This should make it easier to
use the data services (they're now static and no injection is required)
and it should also make it far easier to extend them, either with new
JSON files or with new fields in the model.

Closes #42