Swap logging middleware for enricher (#93)
The middleware added an extra step in the pipeline and only seemed to
work about half of the time -- legitimate identified requests sometimes
had the property set to 'anonymous'.

This approach is probably the correct way of doing it using Serilog
itself rather than a middleware hack, and won't add it if it doesn't get
a value.