Fix infinite session loop (#100)
If your session expired, you get kicked out of
SessionAuthenticationHandler with a SessionException to get a new token.
Problem is that /tool/auth has to go through there too. So you would
never be able to get into AuthService to make a new session. Fix that by
allowing SessionErrors for AllowAnonymous controllers.