Fix savefile import DbConcurrencyException (#102)
If a save import took a long time, the client would re-call `/tool/auth`
again before receiving a response.. This would trigger another save
import while the existing one was running and lead to a
DbConcurrencyException.

This PR:

- Adds a locking mechanism by setting a string derived from the account
ID in Redis
- If a subsequent request is received, and a matching ID string is in
Redis, no import is attempted and the Task is simply to wait until this
key disappears
- Also converts `BaasRequestHelper` to use Redis to cache the jwks, thus
making the main application truly stateless again.