Fix issue with extra units being drawn
- Previous commit reverted HeroParam list being trimmed to MemberCount
when raising CharacterData event. Turns out we need to do this as every
unit sent will be drawn.
- But, the amount of units each player has is worked out based on the
number of HeroParam we get back!
- Reorder the execution of events, so that we first request the
HeroParam, store it for later along with the count in the actor props,
then build the member count table, store it, send the Party event,
and *then* send the CharacterData event based on the MemberCount worked
out during the Party event.