Add 'preferred team' feature (#307)
Implements:

- /quest/get_quest_clear_party
- /quest/get_quest_clear_party_multi
- /quest/set_quest_clear_party
- /quest/set_quest_clear_party_multi

Includes a new database table for saved clear units, and also logic to
detect when entities from saved parties are missing.

Saved clear parties are exempted from the save import deletion process,
as they are not contained in a save, so wiping them on every import
seemed harsh. As a result, they may contain entities which the user no
longer owns.

Fortunately, the game provides a mechanism to handle this in the form of
"lost_unit_list", which shows a pop-up indicating what entities cannot
be retrieved from the clear party.
This is probably intended for entities which can normally be lost, like
dragons or portrait prints, but seems to work sort of okay with weapons
and wyrmprints:

<img
src="https://github.com/SapiensAnatis/Dawnshard/assets/5047192/f769492b-ced9-4fe5-81a0-651f8f161267"
width="30%" height="30%"/>

There is however no entity type for missing shared skills, and missing
characters will cause the button to be disabled entirely.