Fix co-op retry bugs (#371)
Closes #361 and closes #362 

Also move SetGameState call above `base.OnLeave` call to prevent `Plugin
GluonPlugin tries to set game state after call to 'Continue'` error.

---
 
The root cause of both issues was the `StartQuest` property not being
cleared between quest attempts, causing an early start and the proper
data to not be received. This is a plugin-only property we use to decide
when all players are loaded and to fire the StartQuest event.

However, when clearing `StartQuest` on a quest retry, the host appears
to progress through the `GoToIngameState` values _much_ faster than
other players. This leads to other players experiencing an infinite
load, as they never receive the required events -- which probably lead
to #362 since by not clearing `StartQuest` they would be forced to load
into the game without this info.

~~To solve this, the events are now broadcast after a 3-second delay.
This is a bit of a hack, maybe it would be better to find some way to
only send the `GoToIngameState` progression events on the last player
reaching a particular value, but I couldn't find a nice way to do
this.~~

Changed approach: uses an instance variable now to track the min
GoToIngameState value. No need to use `Interlocked` methods as plugin
callbacks are thread-safe.