Fix player identity not being in log context
The changes to service config in #402 forgot to add the log enrichers. I've been meaning to change this to a middleware for a while though, since this way the log properties are only added once per request as opposed to every log.

The PodNameEnricher will not be re-added yet as it is not useful until we start running replicas