Make DbTimeAttackClear key unique for retries (#432)
On a retry in a co-op room, a time attack record request would be
submitted with the same RoomName. The API would think this was the same
room name as a previous clear and fail to register it as a new clear,
instead attempting to update the units. Fix that by incorporating the
room ID -- which is regenerated on retries and clears -- into the unique
ID.