Fix quest cache lookup failure on claiming partial bonus (#451)
When a player skip tickets a quest and claims e.g. 1/3 bonuses, we need
to clear the QuestBonusReserveCount, otherwise they will be asked to
claim the remaining bonuses when they next login, but the cache entry
will be gone.

Also allow the function to return on cache lookup failure, so that a
missing cache entry doesn't block players from logging in.

Closes #447