Don't throw on chara / dragon story unlock failure
Mana node unlocks are regularly failing because players already have all the stories of the character they are trying to unlock. This could be due to a save editor bug. Change this so that it logs a warning instead of throwing; if the story is already unlocked it is probably fine to do nothing even if it shouldn't happen.