Allow quests to drop arbitrary entities (#487)
Closes #397

This refactors the quest drop system to take in a list of entities
rather than just materials, therefore enabling quests to drop Sinister
Dominion wyrmprints.

The quantity algorithm has been changed again, quests now have a drop
table and sample this randomly to generate their drops. The drops are
generated at the start of a given quest and then distributed amongst
enemies, so quests with a lot of enemies will no longer give absurd
quantities of materials.

The downside is that all items now drop as a random amount around the
following averages:

- 50 per clear for materials
- 5 per clear for sindom prints
- 5 per clear for consumables
- 1 per clear for dragon gifts
- 0.1 per clear for other wyrmprints

However we have the capability to adjust these manually without worrying
about the side effects involved with tweaking the entire algorithm. Will
need to supplement this with real data.