Change all primary keys to ViewerId (#500)
Currently all of the tables use DeviceAccountId as a primary key which
has no fixed length and is non-sequential.

Speculatively, this can make joins slower since comparing a string is
likely to be slower than an int, but reading online it seems unlikely
there will be a noticeable performance impact for a DB the size of
Dawnshard's.

It should however improve the organisation of data and make manual
querying / debugging easier