Fix save import error (#521)
The SaveChanges was throwing because the change tracker had stuff in it
which was removed by ExecuteDelete

This highlighted a gap in the save import test coverage, where it was
primarily being tested from the dedicated endpoint which is rarely used.
So an automated test of a save import from /tool/auth has also been
added.