Fix event user data always returning null for events with no items
A check was implemented in #526 to return null for event user data if
the event item dictionary was empty. This was so that the client could
be redirected to the appropriate /entry endpoint. However, this would
lead to infinite loops of calling /entry when the event was meant to
have 0 event items such as Fractured Futures does.

Change the check to be against PlayerEventData which all events are
guaranteed to initialize.