Add data protection key table (#550)
The blazor changes in #536 included enabling antiforgery / csrf tokens
because it was required by the new template. Ever since, the app has
been throwing errors about being unable to validate them because the
keys are not persisted.

Adds a new table using the EF data protection package to store the keys.