Add support for random matching quests to plugin (#551)
For these quests, players are silently placed in a room behind a loading
screen. Photon matchmaking will build up rooms of 1 or more players.
Currently these never start; it seems to be on the plugin to trigger the
start of the room.

- Detect random matching on room creation.
- If we are in random matching, don't send requests to the Redis API, as
these rooms should not be joinable externally.
- If we are in random matching, and a second player joins, set a timer
after which the game will start. The duration is configurable in the
plugin config.
- If the timer is started, and a player leaves such that there is now
only one player in the room, abort the timer.