Fix event item IDs sometimes not matching the item type (#556)
When creating event items, we do a .Zip() of the IDs from
`<type>EventItem` from the MasterAsset into
`Enum.GetValues<{Type}EventItemType>()`.

However, it appears that `data.GetEventSpecificItemIds()` is not
guaranteed to be returned in order, leading to situations like this:


![image](https://github.com/SapiensAnatis/Dawnshard/assets/5047192/54248930-69c3-4785-882c-bb568825c275)

Here, the Id and Type columns are out of sync, leading to event points
being rewarded as gold event items and event points never getting saved.

This PR changes the `.GetEventSpecificItemIds()` to return the
corresponding item type, since it is defined in the same MasterAsset
entry.

I could equally have just added sorting to the returned event item IDs
but using the type defined in the same MasterAsset entry feels like the
safest approach.