Don't use total value for Clear 2 Void Battles mission (#565)
The total value of the EventGroupCleared mission is always 1, so this
mission was impossible to complete if the initial progress handler did
not complete it for you.

It's debatable whether 'Clear 2 Void Battles' means clear 2 _different_
void battles. I have interpreted it in the same way as the 'Clear Five
Quests' daily endeavour which is to make it so 2 clears of the same
quest count.