Make test DB seeding per-method rather than per-class (#568)
Moves the creation of the test savefile and respawning (clearing) of the
database into `TestFixture`, which is instantiated once for every test
method. It is currently in `CustomWebApplicationFactory`, which runs
once per test class. This often introduces conflicts between tests as
running one test after another causes the second test to run upon the
mutated state from the first test.

With these changes, each test method operates on a clean set of data
which should hopefully make random failures less frequent.

In addition:
- Implements a
[CollectionFixture](https://xunit.net/docs/shared-context#collection-fixture)
which controls the Postgres and Redis containers, rather than using a
static `TestContainers` class which indirectly runs its setup when first
accessed.
- Moves a lot of setup to use actual async in
`IAsyncLifetime.InitializeAsync` over calling `.Wait()` on async methods
in constructors. This has the side effect that setup in test class
constructors cannot now act on the savefile since `InitializeAsync` is
run after the constructor.
- Implements a `Task Setup()` method which each test class can override
for async setup.
- Disables query tracking for the instance of `ApiContext` exposed by
`TestFixture`; because this instance is primarily used for querying and
change tracking can produce odd results for 'before and after'
comparisons. This can be re-enabled by calling `AsTracking` on a query.