Configure workflows to run on main
I'm going to delete the `master` branch and rename `develop` to `main`.

We don't get a lot of use out of `master` because we don't do stable releases -- we have done in the past but these have never been deployed, and were more for the benefit of versioning the photon plugin / state manager.