Use System.Text.Json in plugin over Newtonsoft (#665)
Thinking about the build issue in #664 and how adding STJ fixed it
(still not sure why) I decided it's probably preferable to use STJ over
Newtonsoft, as then we can remove it from the solution.

Replaces Netwonsoft with STJ in the plugin. Retarget to net481 (latest
netfx release) after anecdotal evidence from Photon discord that this
works OK. This should fix most of our woes with packages not being
supported