Fix a few occurrences of KeyNotFoundException in the plugin (#666)
Most of these are to do with the logic to bring the minimum
GoToIngameState forward when an actor leaves. Typically the cause is
iterating through `PluginHost.GameActors` and then indexing each
element's `ActorNr` into a state dictionary... we remove the actor from
the state dictionary on leave, but the actor is not always removed from
the plugin host list at the same time.

Generally fixed by looping through the dictionaries instead of
`PluginHost.GameActors` in `RaisePartyEvent` and
`RaiseCharacterDataEvent`

Additionally:
- Refactor all `GoToIngameState` logic into `GoToIngameStateManager` to
make `GameLogicPlugin` more manageable
- Change the raising of the character data event into `GoToIngameState =
4`, based on the enum in the il2cpp.dll specifying this value:

  ```cs
  public enum GoToIngameState
  {
	  None,
	  TryGoToIngame,
	  PreCheckParty,
	  CheckParty,
	  PostCheckParty
  }
  ```

NOTE: for leave logic to become effective, adjust
`MinimumTimeout`/`MaximumTimeout` down in bin_Win64/PhotonServer.config