Remove empty request models
Lots of MessagePack deserialization errors in the logs after #676.

MessagePack.MessagePackSerializationException: Failed to deserialize DragaliaAPI.Models.Generated.DragonGetContactDataRequest value.
 ---> MessagePack.MessagePackSerializationException: Unexpected msgpack code 128 (fixmap) encountered.

The common thread between them seems to be on endpoints with empty request
bodies where we are trying to bind said empty request body. Possibly the
fact that we are no longer setting keysAsPropertyNames = true is making
MessagePack-CSharp think we expect an array of integer keys (as we have no properties to specify string keys on), but the game
is still sending a map.