Stop sending summon ticket info
This is currently blocking summoning with payment target errors, but
the way the data is currently being stored isn't quite right. If we
disabled the payment target, the next error would be material shortage
since we store each ticket with its own key ID and don't delete them
when the quantity reaches 0.

Needs a rethink - so for now disable sending tickets so that people
can still summon with wyrmite.