Fix summoning ticket data storage and re-enable usage (#690)
Add a savefile update to consolidate stackable summoning ticket
quantities and change the way they are rewarded, so that we can use them
in the summoning endpoint.

Previously, each time you got a summoning ticket, it was a new row with
its own key ID and quantity. So when the payment service came to deduct
a ticket, it would not necessarily see all of the tickets you had.
Judging by the sample endgame savefile, the game did not distribute
stackable tickets across multiple key IDs. I assume the key ID column
was designed for expiring tickets? Evidence from other saves suggests
special tickets could also be stacked.

Also removes SummonTicketRepository and enables global query filters for
getting summon tickets. This is a trial to see if we can abandon the
repository pattern in favour of injecting ApiContext directly.