Fix "invalid repeat key" warning spam
This appears to be caused by the client sending a stale repeat key, instead of null, if you auto repeat a quest and then auto repeat another quest after ending the first auto repeat.

Change the handling so that when we receive a repeat key which is invalid, we act in the same way as if we had received null on a first clear, and go to the viewer ID lookup.