Handle story PK conflict when adding characters (#718)
When adding a character, if it is not already owned by the player a
story will always be added. This is not a problem for normal gameplay
logic, but several players have run into issues after removing
characters from their imported saves and leaving their stories behind.

Fix this by adding an extra check for a story in the UnitRepository.
Also misc. tidying up since that code is oold