Fix Harle / Origa being unobtainable (#748)
The GuestJoinStoryId for these events in the StoryData asset is
inaccurate and refers to a story from outside of the event, so our
mechanism to award them will never trigger.

Adds a custom overrides table to rectify this - these events appear to
be the only ones affected based on the below query:

```sql
SELECT _Id FROM "EventData" 
WHERE _GuestJoinStoryId / 100 <> _Id AND _GuestJoinStoryId <> 0 AND _IsMemoryEvent = 1;
```

To-do:
- [x] Tests
- [x] Savefile update

Also did a bit of refactoring in our quest to remove repositories -
added a global query filter to DbPlayerStoryState and DbPlayerCharaData
i.e. the tables needed for the new savefile update. This also included
refactoring PartyController a bit since I saw the validation code and
almost cried at the looped db query.