Fix dragon's roost gifting softlock (#750)
Fix a softlock when reaching bond level 5 or 15 in the Dragon's Roost.
We were creating an array of size 1 but early returning if the bond
level unlocked a story. This lead to returning a list of `[null]`, which
the game didn't like. This PR changes that to return an empty list `[]`
in this case.

It also makes it so that the reward.IsReleaseStory flag is set even if
the story fails to unlock due to invalid savedata. This is because if
you send `[]` without `IsReleaseStory` then it shows an empty gifts
screen, which is a minor visual glitch.