Fix weapon passives not applying when using off-element weapons (#752)
The current logic to retrieve weapon passive abilities uses the equipped
weapon's element, when it should use the character's element. This
manifests as not having any passive abilities active if you use an
off-element weapon, because the off-element passive abilities won't
apply correctly. The actual behaviour should be that a weapon passive
that has been unlocked is active for all units of that element and
weapon type, regardless of what weapon is being used.

Additionally removes if statements that only send weapon passive data if
a weapon is equipped. Weapon passive abilities should be available even
when no weapon is equipped.