Maybe fix NullReferenceException in party loading code
I've observed the following exception in the logs since #752 was
merged:

```
System.NullReferenceException: Object reference not set to an instance of an object.
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartService.ProcessDetailedUnitList(List`1 detailedPartyUnits) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartService.cs:line 292
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartService.GetAssignUnitIngameData(Int32 questId, IList`1 party, Nullable`1 supportViewerId, RepeatSetting repeatSetting) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartService.cs:line 151
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartController.StartAssignUnit(DungeonStartStartAssignUnitRequest request, CancellationToken cancellationToken) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartController.cs:line 105
```

Nobody has come forward with a report or reproduction steps and I can't
seem to trigger it myself - I had thought it might be caused by having
empty unit slots but this is not the case. This is a shot in the dark at
fixing it. If it doesn't work, then there's some additional logging
included which should help to get to the bottom of the issue.