Maybe fix NullReferenceException in party loading code  (#755)
I've observed the following exception in the logs since
https://github.com/SapiensAnatis/Dawnshard/pull/752 was merged:

```
System.NullReferenceException: Object reference not set to an instance of an object.
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartService.ProcessDetailedUnitList(List`1 detailedPartyUnits) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartService.cs:line 292
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartService.GetAssignUnitIngameData(Int32 questId, IList`1 party, Nullable`1 supportViewerId, RepeatSetting repeatSetting) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartService.cs:line 151
   at DragaliaAPI.Features.Dungeon.Start.DungeonStartController.StartAssignUnit(DungeonStartStartAssignUnitRequest request, CancellationToken cancellationToken) in /src/DragaliaAPI/DragaliaAPI/Features/Dungeon/Start/DungeonStartController.cs:line 105
```

Nobody has come forward with a report or reproduction steps and I can't
seem to trigger it myself. Since it started happening after we used
`CharaData` to load up weapon passives, and because this is not in a
null guard, I assume `detailedUnit.CharaData` is somehow null, so I've
added a null guard.

If this doesn't work, then there's some additional logging included
which should help to get to the bottom of the issue.