Fix AddStamina not including passive recovery (#774)
Ensure that the AddStamina method updates the current stamina before it
adds the new stamina on to the user.

Fixes an issue with honey tea where sometimes claiming it could visually
reduce your stamina - the interface shows the 'projected' stamina
including passive recovery, and by updating it based on the underlying
un-updated stamina value, we would set it to the wrong value.