New website: homepage (#789)
I am looking to rewrite the existing webpage in a JavaScript framework
for a few reasons:

- Blazor Server requires a constant websocket connection, which can be
prone to dropping.
- The original in-game webviews used JavaScript and assorted tricks such
as manipulating the window hash (see #494). It will be easier to achieve
parity with a JavaScript framework.
- I believe the depth of customization and the ecosystem around JS apps
will be needed if we want to pursue a custom component library for fully
replicating the style of the in-game webviews.

Point 1 and point 2 to a degree could be counteracted using Blazor
WebAssembly, but the initial load times on this are unacceptably long
especially for the in-game webviews.

I've chosen Svelte because I wanted to try and learn it and enjoyed what
I saw. I think that React would add too much complexity for a website
that ultimately does not need to do a lot.

This PR contains the homepage, which just sets outs the projects aims,
status and has a how-to-play guide. It won't currently deploy anywhere
and there is no CI. I am just making a pull request now so that I don't
have to make one when the rewrite is 'done' with 1000+ changed files.