Add limit for dragons (#792)
Starts enforcing the maximum dragon limit, with excess going to the
present box. Closes
https://github.com/SapiensAnatis/Dawnshard/issues/670

On save import, the dragon list will be trimmed, taking the first N
dragons according to the `$.user_data.max_dragon_quantity`. This itself
is capped at 500 to prevent people from editing it to `int.MaxValue`.
Existing saves will not be affected unless they are re-imported.

Additionally includes some refactoring of `RewardService` and the
summoning logic. The existing 'grant multiple character/dragon' logic
has been extracted into `RewardService` as an implementation of a new
`IBatchRewardHandler`. `SummonService` now calls this from a support
`UnitService` that handles the unique requirement of needing to know
whether an added dragon was new, since this cannot be known from the
reward grant result.