Delete IResetHelper and replace with extension method (#795)
Removes the IResetHelper interface / class and replaces it with a set of
extension methods on TimeProvider. This is a bit of refactoring I had in
the back of my mind, which I now want to do ahead of adding MG reward
popups which would have used IResetHelper.

IResetHelper is not really a beneficial abstraction because the reset
_should_ be explicitly coupled to the current time. It does not make
sense to explicitly mock the last reset, rather you should mock the
current time.

Also snuck in some changes to:
  - Move GlobalUsings into csproj
  - Remove unneeded usings solution-wide (partly to assist with errors
after the DragaliaAPI.Helpers namespace was nuked)