Add warning on save export page when exceeding dragon limit
This is based on player feedback from the changes to save import in
#792 to provide a more clear warning about what will happen when
importing a save with an excess number of dragons. Since we are not
currently able to show feedback on the save import process, we will
settle for introducing a dialog in the save export page where most
players who end up reimporting their save will have to go first.