Allow claiming Mercurial Gauntlet monthly rewards (#797)
Adds a check on `/login/index` to prompt the client to claim mercurial
gauntlet rewards if the player is eligible, as tracked by a new table
showing the last reward date.

The last reward date row is populated in InitializeWall(), and for
players currently playing in a new V21Update.

Also:

- Delete WallRepository
- Delete LoginBonusRepository
- Add twinkling sands to the rewards
- Add some additional validation to `/wall/receive_monthly_reward`
- Flesh out `/wall/get_monthly_reward`