Publish statemanager events for private rooms
The practice of not publishing StateManager events when a room's
ShouldPublish flag is false is probably incorrect, since it breaks the
ability to join private rooms via passcode -- no room is ever created
in Redis.

In general, it is probably best to update StateManager in all cases
even if not _strictly_ necessary, for example updating player count in
random matching rooms. The requests are generally non-blocking and will
not cause any user-facing announcements. So we should limit the publish
restrictions to the Discord plugin which does make user-facing
announcements.