Fix T3 event materials being unavailable from certain events
Closes #507

Replaces the calculation based on VariationType with one that looks at
the least significant digits of the quest ID, as these always follow a
particular structure for wave quests across all events. It produces an
identical result to the old calculation assuming that the old
calculation was intended for typical events with VariationType 3 and 4
in the wave quests.

+ `launchSettings.json` fix for making the server accessible via a
mobile device